+++ src/routes/jira/connect/enterprise/jira-connect-enterprise-router.ts
@@ -3,7 +3,7 @@ import { JiraJwtTokenMiddleware } from "middleware/jira-jwt-middleware";
 import { csrfMiddleware } from "middleware/csrf-middleware";
 import { JiraConnectEnterpriseGet } from "./jira-connect-enterprise-get";
 import { JiraConnectEnterprisePost } from "routes/jira/connect/enterprise/jira-connect-enterprise-post";
-import { JiraConnectEnterpriseServerAppRouter } from "routes/jira/connect/enterprise/app-for-server/jira-connect-enterprise-server-app-router";
+import { JiraConnectEnterpriseServerAppRouter } from "routes/jira/connect/enterprise/server_app/jira-connect-enterprise-server-app-router";
 import { JiraConnectEnterpriseAppRouter } from "routes/jira/connect/enterprise/app/jira-connect-enterprise-app-router";
 
 export const JiraConnectEnterpriseRouter = Router();
