+++ src/routes/jira/connect/enterprise/server_app/jira-connect-enterprise-server-app-router.ts
@@ -1,7 +1,7 @@
 import { Router } from "express";
 import { JiraJwtTokenMiddleware } from "middleware/jira-jwt-middleware";
 import { csrfMiddleware } from "middleware/csrf-middleware";
-import { JiraConnectEnterpriseServerAppGet } from "routes/jira/connect/enterprise/app-for-server/jira-connect-enterprise-server-app-get";
+import { JiraConnectEnterpriseServerAppGet } from "routes/jira/connect/enterprise/server_app/jira-connect-enterprise-server-app-get";
 import { JiraConnectEnterpriseAppCreateOrEdit } from "routes/jira/connect/enterprise/app/jira-connect-enterprise-app-create-or-edit";
 
 export const JiraConnectEnterpriseServerAppRouter = Router();
